split oc3 in 2 files
