+++ chains/solana/contracts/programs/ccip-router/src/instructions/v1/mod.rs
@@ -12,5 +12,6 @@ pub mod token_admin_registry;
 mod fee_quoter;
 mod merkle;
 mod messages;
-mod ocr3;
+mod ocr3base;
+mod ocr3impl;
 mod pools;
