+++ chains/solana/contracts/programs/ccip-router/src/instructions/v1/ocr3base.rs
@@ -2,7 +2,7 @@ use anchor_lang::prelude::*;
 use anchor_lang::solana_program::sysvar;
 use anchor_lang::solana_program::{keccak, secp256k1_recover::*};
 
-use crate::{CommitInput, ExecutionReportSingleChain, Ocr3Config, Ocr3ConfigInfo};
+use crate::{Ocr3Config, Ocr3ConfigInfo};
 
 #[constant]
 pub const MAX_ORACLES: usize = 16; // can set a maximum of 16 transmitters + 16 signers simultaneously in a single set config tx
@@ -111,6 +111,11 @@ fn clear_signers(ocr3_config: &mut Ocr3Config) {
 
 fn clear_transmitters(ocr3_config: &mut Ocr3Config) {
     ocr3_config.transmitters = [[0; 32]; MAX_TRANSMITTERS]
+}
+
+pub trait Ocr3Report {
+    fn hash(&self, ctx: &ReportContext) -> [u8; 32];
+    fn len(&self) -> usize;
 }
 
 pub fn ocr3_transmit<R: Ocr3Report>(
@@ -224,52 +229,6 @@ fn assign_oracles<const A: usize>(oracles: &mut [[u8; A]], location: &mut [[u8; 
         location[i] = *o;
     }
     Ok(())
-}
-
-pub trait Ocr3Report {
-    fn hash(&self, ctx: &ReportContext) -> [u8; 32];
-    fn len(&self) -> usize;
-}
-
-pub struct Ocr3ReportForCommit<'a>(pub &'a CommitInput);
-
-impl Ocr3Report for Ocr3ReportForCommit<'_> {
-    fn hash(&self, ctx: &ReportContext) -> [u8; 32] {
-        use anchor_lang::solana_program::hash;
-        let mut buffer: Vec<u8> = Vec::new();
-        self.0.serialize(&mut buffer).unwrap();
-        let report_len = self.len() as u16; // u16 > max tx size, u8 may have overflow
-        hash::hashv(&[&report_len.to_le_bytes(), &buffer, &ctx.as_bytes()]).to_bytes()
-    }
-
-    fn len(&self) -> usize {
-        4 + (32 + 28) * self.0.price_updates.token_price_updates.len() + // token_price_updates
-        4 + (8 + 28) * self.0.price_updates.gas_price_updates.len() + // gas_price_updates
-        self.0.merkle_root.len()
-        // + 4 + 65 * self.rmn_signatures.len()
-    }
-}
-
-pub struct Ocr3ReportForExecutionReportSingleChain<'a>(pub &'a ExecutionReportSingleChain);
-
-impl Ocr3Report for Ocr3ReportForExecutionReportSingleChain<'_> {
-    fn hash(&self, _: &ReportContext) -> [u8; 32] {
-        [0; 32] // not needed, this report is not hashed for signing
-    }
-    fn len(&self) -> usize {
-        let offchain_token_data_len = self
-            .0
-            .offchain_token_data
-            .iter()
-            .fold(0, |acc, e| acc + 4 + e.len());
-
-        8 // source chain selector
-        + self.0.message.len() // ccip message
-        + 4 + offchain_token_data_len// offchain_token_data
-        + 32 // root
-        + 4 + self.0.proofs.len() * 32 // count + proofs
-        + 4 + self.0.token_indexes.len() // token_indexes
-    }
 }
 
 #[error_code]
